requirements.txt
pandas
matplotlib
seaborn
streamlit
crewai
crewai-tools
python-dotenv
litellm
reportlab
openpyxl